Move types in navigator
The project_tree.go file was getting too messy to navigate easily.  It's better now.
